Convert ConsoleCommand to use generic message
